requirements.txt
Django==2.2.13
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1
django-heroku==0.3.1
Pillow==7.1.2
flake8==3.8.3
django-admin-rangefilter==0.6.1
django-storages==1.10
boto3==1.11.7
django-extensions==2.2.9
whitenoise==5.1.0
sorl-thumbnail==12.6.3
mercadopago==1.1.1
random-username==1.0.2
WeasyPrint==51